+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- Ensure `requests.Session` is closed to prevent lingering TCP connections.
 
 ## [0.4.0] - 2025-01-20
 ### Changed